Add: introduce more clear guidelines in the use of GPL licenses
Given the following scenario:
- Author A creates a GRF under GPL v2
- Author B picks up that work and extends on it
If user B wants to upload to BaNaNaS, the only way he can do that
is by complying to GPL v2. In order to be compliant, a link to
the source code has to be supplied. This commit makes it more
clear to the authors that this is expected of them.

This is not meant to protect OpenTTD, but more to be a good
neighbour and ensure authors are aware of the implications of
the license they used (to base their work on).